requirements.txt
beautifulsoup4
requests
pymongo[srv]==python_version
python-dotenv
scrapy